+++ src/Client/dotnet-graphql/Command.cs
@@ -1,0 +1,8 @@
+namespace StrawberryShake.Tools
+{
+    public enum Command
+    {
+        Init,
+        Compile
+    }
+}
