+++ src/Client/dotnet-graphql/Configuration.cs
@@ -1,0 +1,10 @@
+using System.Collections.Generic;
+
+namespace StrawberryShake.Tools
+{
+    public class Configuration
+    {
+        public List<SchemaFile> Schemas { get; set; }
+        public string ClientName { get; set; }
+    }
+}
