+++ src/Client/dotnet-graphql/ICommand.cs
@@ -1,0 +1,9 @@
+using System.Threading.Tasks;
+
+namespace StrawberryShake.Tools
+{
+    public interface ICommand
+    {
+        Task<int> OnExecute();
+    }
+}
