+++ src/Client/dotnet-graphql/SchemaFile.cs
@@ -1,0 +1,9 @@
+namespace StrawberryShake.Tools
+{
+    public class SchemaFile
+    {
+        public string Name { get; set; }
+        public string Type { get; set; }
+        public string File { get; set; }
+    }
+}
