+++ src/Client/dotnet-graphql/dotnet-graphql.csproj
@@ -3,9 +3,9 @@
   <PropertyGroup>
     <OutputType>Exe</OutputType>
     <TargetFramework Condition="'$(TargetFrameworks)'==''">netcoreapp2.2</TargetFramework>
-    <RootNamespace>StraberryShake.Tools</RootNamespace>
-    <PackageId>StraberryShake.Tools</PackageId>
+    <PackageId>StrawberryShake.Tools</PackageId>
     <AssemblyName>dotnet-graphql</AssemblyName>
+    <RootNamespace>StrawberryShake.Tools</RootNamespace>
     <IsPackable>true</IsPackable>
     <PackAsTool>true</PackAsTool>
   </PropertyGroup>
